backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
openai
python-dotenv
